auth_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
google-auth
python-multipart
python-dotenv
